Code Cleanup & Shader Playing
Learned how Triplanar works (at least a decent implimentation, I knda already assumed how it worked by sampling 3 times but never knew how to properly blend it)

+++ Assets/Scripts/UniVox/VoxelData/World.cs
@@ -4,7 +4,7 @@ using System.Collections.Generic;
 using Unity.Entities;
 using Unity.Mathematics;
 
-namespace UniVox.Core.Types
+namespace UniVox.VoxelData
 {
     public class World : IDisposable, IReadOnlyDictionary<int3, Entity>
     {
@@ -51,8 +51,12 @@ namespace UniVox.Core.Types
             return Records.TryGetValue(key, out accessor);
         }
 
+        private bool disposed;
         public void Dispose()
         {
+            if(disposed)
+                return;
+            disposed = true;
             foreach (var recordValue in Records.Values)
             {
                 EntityManager.DestroyEntity(recordValue);
